requirements.txt
Django>=5.0,<6
openai>=1.3.0
python-dotenv>=1.0.0
supabase>=2.6.0
